+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changes in Version 3.3.1
 ------------------------
 * Changes the exception isinstance check to look for basestring/str instead of
   bytes in order to properly store the error description.
+* Fixed error where you could not create a LUN with the ID of 0. (issue #17)
 
 Changes in Version 3.3.0
 ------------------------